backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pypdf==5.1.0
python-pptx==1.0.2
python-dotenv==1.0.1
langchain==0.3.13
langchain-community==0.3.13
faiss-cpu==1.9.0
tiktoken==0.8.0
openai-whisper==20231117
sentence-transformers==2.2.2
langdetect==1.0.9
langchain-groq==0.2.1